# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCConnect API",
    "version": "2.0.0",
    "description": "Authoritative REST contract for PCConnect controllers and agents. SignalR notifications are advisory; REST resources are authoritative."
  },
  "servers": [
    { "url": "https://api.pcconnect.adamdeveloping.co.uk/api/v2", "description": "Production" }
  ],
...
  "paths": {
...
    "/reminders": {
      "get": {
        "tags": ["Reminders"], "operationId": "listReminders", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "description": "Reminders", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReminderPage" } } } } }
      },
      "post": {
        "tags": ["Reminders"], "operationId": "createReminder", "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReminderWrite" } } } },
        "responses": { "201": { "description": "Reminder created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reminder" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/reminders/{reminderId}": {
      "get": { "tags": ["Reminders"], "operationId": "getReminder", "parameters": [{ "$ref": "#/components/parameters/ReminderId" }], "responses": { "200": { "description": "Reminder", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reminder" } } } }, "404": { "$ref": "#/components/responses/Problem" } } },
      "patch": {
        "tags": ["Reminders"], "operationId": "updateReminder", "parameters": [{ "$ref": "#/components/parameters/ReminderId" }],
        "requestBody": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/ReminderWrite" } } } },
        "responses": { "200": { "description": "Updated reminder", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Reminder" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      },
      "delete": { "tags": ["Reminders"], "operationId": "deleteReminder", "parameters": [{ "$ref": "#/components/parameters/ReminderId" }], "responses": { "204": { "description": "Reminder cancelled and deleted" }, "404": { "$ref": "#/components/responses/Problem" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque" }
    },
    "parameters": {
      "Cursor": { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "maxLength": 512 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
...
      "ReminderId": { "name": "reminderId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
    },
    "responses": {
      "Problem": { "description": "RFC 9457 problem details", "headers": { "X-Correlation-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Instant": { "type": "string", "format": "date-time" },
...
      "Problem": { "type": "object", "additionalProperties": true, "required": ["type", "title", "status", "code", "correlationId"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "detail": { "type": "string" }, "instance": { "type": "string", "format": "uri-reference" }, "code": { "type": "string", "pattern": "^[a-z0-9_]+$" }, "correlationId": { "type": "string" }, "errors": { "type": "array", "items": { "type": "object", "required": ["pointer", "code"], "properties": { "pointer": { "type": "string" }, "code": { "type": "string" } } } } } },
...
      "ReminderTargetMode": { "type": "string", "enum": ["all_devices", "selected_devices"] },
      "ReminderWrite": { "type": "object", "additionalProperties": false, "required": ["text", "targetMode", "timezone", "localStart"], "properties": { "text": { "type": "string", "minLength": 1, "maxLength": 2000 }, "targetMode": { "$ref": "#/components/schemas/ReminderTargetMode" }, "targetDeviceIds": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/Uuid" } }, "timezone": { "type": "string", "minLength": 1, "maxLength": 100 }, "localStart": { "type": "string", "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:\\.[0-9]+)?$", "description": "Local date-time without offset interpreted in timezone." }, "recurrenceRule": { "type": ["string", "null"], "maxLength": 1000, "description": "RFC 5545 RRULE without DTSTART." }, "expectedVersion": { "type": "integer", "minimum": 1 } }, "allOf": [{ "if": { "properties": { "targetMode": { "const": "selected_devices" } }, "required": ["targetMode"] }, "then": { "required": ["targetDeviceIds"], "properties": { "targetDeviceIds": { "minItems": 1 } } } }] },
      "Reminder": { "type": "object", "additionalProperties": false, "required": ["id", "text", "targetMode", "targetDeviceIds", "timezone", "localStart", "createdAt", "version"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "text": { "type": "string" }, "targetMode": { "$ref": "#/components/schemas/ReminderTargetMode" }, "targetDeviceIds": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/Uuid" } }, "timezone": { "type": "string" }, "timezoneAssumed": { "type": "boolean" }, "localStart": { "type": "string", "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:\\.[0-9]+)?$" }, "recurrenceRule": { "type": ["string", "null"] }, "nextOccurrenceAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "version": { "type": "integer", "minimum": 1 } } },
      "ReminderPage": { "allOf": [{ "$ref": "#/components/schemas/Page" }, { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Reminder" } } } }] },
...
      "Page": { "type": "object", "additionalProperties": false, "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": true }, "nextCursor": { "type": ["string", "null"] } } }
    }
  },
  "security": [{ "bearerAuth": [] }]
}